+++ app/code/Magento/Adminhtml/Block/Promo/Quote/Edit/Tab/Coupons/Form.php
@@ -158,6 +158,6 @@ class Form
      */
     public function getGenerateUrl()
     {
-        return $this->getUrl('*/*/generate');
+        return $this->getUrl('adminhtml/*/generate');
     }
 }