+++ app/code/Magento/Sales/Block/Adminhtml/Order/Create/Coupons.php
@@ -1,0 +1,44 @@
+<?php
+/**
+ * {license_notice}
+ *
+ * @category    Magento
+ * @package     Magento_Sales
+ * @copyright   {copyright}
+ * @license     {license_link}
+ */
+
+/**
+ * Adminhtml sales order create coupons block
+ *
+ * @category   Magento
+ * @package    Magento_Sales
+ * @author      Magento Core Team <core@magentocommerce.com>
+ */
+
+namespace Magento\Sales\Block\Adminhtml\Order\Create;
+
+class Coupons extends \Magento\Sales\Block\Adminhtml\Order\Create\AbstractCreate
+{
+
+    protected function _construct()
+    {
+        parent::_construct();
+        $this->setId('sales_order_create_coupons_form');
+    }
+
+    public function getCouponCode()
+    {
+        return $this->getQuote()->getCouponCode();
+    }
+
+    public function getHeaderText()
+    {
+        return __('Coupons');
+    }
+
+    public function getHeaderCssClass()
+    {
+        return 'head-promo-quote';
+    }
+}