+++ app/code/Magento/Sales/Block/Adminhtml/Order/Create/Newsletter.php
@@ -1,0 +1,45 @@
+<?php
+/**
+ * {license_notice}
+ *
+ * @category    Magento
+ * @package     Magento_Sales
+ * @copyright   {copyright}
+ * @license     {license_link}
+ */
+
+/**
+ * Adminhtml sales order create newsletter block
+ *
+ * @category   Magento
+ * @package    Magento_Sales
+ * @author      Magento Core Team <core@magentocommerce.com>
+ */
+
+namespace Magento\Sales\Block\Adminhtml\Order\Create;
+
+class Newsletter extends \Magento\Sales\Block\Adminhtml\Order\Create\AbstractCreate
+{
+
+    protected function _construct()
+    {
+        parent::_construct();
+        $this->setId('sales_order_create_newsletter');
+    }
+
+    public function getHeaderText()
+    {
+        return __('Newsletter Subscription');
+    }
+
+    public function getHeaderCssClass()
+    {
+        return 'head-newsletter-list';
+    }
+
+    protected function _toHtml()
+    {
+        return parent::_toHtml();
+    }
+
+}