+++ app/code/Magento/Sales/Block/Adminhtml/Order/Create/Search/Grid/Renderer/Price.php
@@ -1,0 +1,37 @@
+<?php
+/**
+ * {license_notice}
+ *
+ * @category    Magento
+ * @package     Magento_Sales
+ * @copyright   {copyright}
+ * @license     {license_link}
+ */
+
+/**
+ * Adminhtml sales create order product search grid price column renderer
+ *
+ * @category   Magento
+ * @package    Magento_Sales
+ * @author      Magento Core Team <core@magentocommerce.com>
+ */
+namespace Magento\Sales\Block\Adminhtml\Order\Create\Search\Grid\Renderer;
+
+class Price extends
+    \Magento\Adminhtml\Block\Widget\Grid\Column\Renderer\Price
+{
+    /**
+     * Render minimal price for downloadable products
+     *
+     * @param   \Magento\Object $row
+     * @return  string
+     */
+    public function render(\Magento\Object $row)
+    {
+        if ($row->getTypeId() == 'downloadable') {
+            $row->setPrice($row->getPrice());
+        }
+        return parent::render($row);
+    }
+
+}