+++ app/code/Magento/Sales/Block/Adminhtml/Order/View/Tabs.php
@@ -1,0 +1,69 @@
+<?php
+/**
+ * {license_notice}
+ *
+ * @category    Magento
+ * @package     Magento_Sales
+ * @copyright   {copyright}
+ * @license     {license_link}
+ */
+
+/**
+ * Order view tabs
+ */
+namespace Magento\Sales\Block\Adminhtml\Order\View;
+
+class Tabs extends \Magento\Backend\Block\Widget\Tabs
+{
+    /**
+     * Core registry
+     *
+     * @var \Magento\Core\Model\Registry
+     */
+    protected $_coreRegistry = null;
+
+    /**
+     * @param \Magento\Core\Helper\Data $coreData
+     * @param \Magento\Backend\Block\Template\Context $context
+     * @param \Magento\Backend\Model\Auth\Session $authSession
+     * @param \Magento\Core\Model\Registry $registry
+     * @param array $data
+     */
+    public function __construct(
+        \Magento\Core\Helper\Data $coreData,
+        \Magento\Backend\Block\Template\Context $context,
+        \Magento\Backend\Model\Auth\Session $authSession,
+        \Magento\Core\Model\Registry $registry,
+        array $data = array()
+    ) {
+        $this->_coreRegistry = $registry;
+        parent::__construct($coreData, $context, $authSession, $data);
+    }
+
+    /**
+     * Retrieve available order
+     *
+     * @return \Magento\Sales\Model\Order
+     */
+    public function getOrder()
+    {
+        if ($this->hasOrder()) {
+            return $this->getData('order');
+        }
+        if ($this->_coreRegistry->registry('current_order')) {
+            return $this->_coreRegistry->registry('current_order');
+        }
+        if ($this->_coreRegistry->registry('order')) {
+            return $this->_coreRegistry->registry('order');
+        }
+        throw new \Magento\Core\Exception(__('We cannot get the order instance.'));
+    }
+
+    protected function _construct()
+    {
+        parent::_construct();
+        $this->setId('sales_order_view_tabs');
+        $this->setDestElementId('sales_order_view');
+        $this->setTitle(__('Order View'));
+    }
+}