+++ dev/tests/functional/tests/app/Magento/Backend/Test/Handler/Ui/LoginUser.php
@@ -1,0 +1,50 @@
+<?php
+/**
+ * {license_notice}
+ *
+ * @spi
+ * @category    Mtf
+ * @package     Mtf
+ * @subpackage  functional_tests
+ * @copyright   {copyright}
+ * @license     {license_link}
+ */
+
+namespace Magento\Backend\Test\Handler\Ui;
+
+use Mtf\Fixture;
+use Mtf\Handler\Ui;
+use Mtf\Factory\Factory;
+
+/**
+ * Class LoginUser
+ * Handler for ui backend user login
+ *
+ * @package Magento\Backend\Test\Handler\Ui
+ */
+class LoginUser extends Ui
+{
+    /**
+     * Login admin user
+     *
+     * @param Fixture $fixture [optional]
+     * @return mixed|string
+     */
+    public function execute(Fixture $fixture = null)
+    {
+        if (null === $fixture) {
+            $fixture = Factory::getFixtureFactory()->getMagentoBackendAdminSuperAdmin();
+        }
+
+        $loginPage = Factory::getPageFactory()->getAdminAuthLogin();
+        $loginForm = $loginPage->getLoginBlockForm();
+
+        $loginPage->open();
+
+        $adminHeaderPanel = $loginPage->getHeaderPanelBlock();
+        if (!$adminHeaderPanel || !($adminHeaderPanel->isVisible())) {
+            $loginForm->fill($fixture);
+            $loginForm->submit();
+        }
+    }
+}