+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-* Removed residual page fragment code
+* Cleavage of Adminhtml module
+  * Move Sales-related logic to Sales module
+  * Move Catalog-related logic to Catalog module
+  * Move Tax-related logic to Tax module
+  * Move: Action, Cache, Ajax, Dashboard, Index, Json, Rating, Sitemap, Survey, UrlRewrite from root of Adminhtml Controller folder
 * View abstraction was moved into lib
 * Templates and layouts are updated Weee, Directory, Sales, Payment, PaypalUk, Paypal modules, old files moved to magento_backup theme
 * Templates and layouts are updated in Cms, Contacts, Core, GoogleCheckout, Page, Rating, Review, Rss, Widget modules, old files moved to magento_backup theme
@@ -10,6 +14,7 @@
 * Fixed Impossible to add image using WYSIWYG
 * No direct code execution in layout: methods addColumnRender, addRenderer, addToParentGroup usages as action nodes were eliminated
 * Legacy static test ObsoleteCodeTest::testPhpFiles produced false-positive results
+* Removed residual page fragment code
 
 2.0.0.0-dev49
 =============